subcmd_dump_packed_instructions.cpp

+++ src/subcmd_dump_packed_instructions.cpp
@@ -185,7 +185,7 @@ namespace l
             unpack_row(row,bs,ccc);
             for(u64 i = offset; i < next_offset; i += 1)
               {
-                u32 x = bs.read(offset,BITS_PER_BYTE);
+                u8 x = bs.read(offset,BITS_PER_BYTE);
                 fmt::print("{:01X} ",x);
               }
             fmt::print("\n");
